requirements.txt
Flask
Flask-Script
gunicorn
click
redis
gevent
pytest
requests
pyyaml
